# app/src/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Error body returned by every failing endpoint.\n\n`status` is the HTTP status [`IntoResponse`] answers with. It is not part of\nthe JSON body — it would be redundant with the response line, and clients\ngenerated from the OpenAPI schema must not see a field the server never\nserializes — so it is `#[serde(skip)]`ped and defaults to 400 on both\nconstruction and deserialization.",
        "properties": {
          "code": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "running_job_id": {
            "description": "The job holding the single execution slot, on the `409` from\n`POST /api/v1/scripts/execute`. Skipped when absent so every other\nerror body is byte-for-byte what it was before this field existed --\na client that switched on the presence of a key would otherwise start\nseeing `\"running_job_id\": null` on unrelated failures.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "message",
          "code"
        ],
        "type": "object"
      },
...
      "RunFramesResponse": {
        "description": "`GET /api/v1/runs/{id}/frames` response.",
        "properties": {
          "frames": {
            "items": {
              "$ref": "#/components/schemas/ArchivedFrame"
            },
            "type": "array"
          }
        },
        "required": [
          "frames"
        ],
        "type": "object"
      },
...
      "RunsResponse": {
        "description": "`GET /api/v1/runs` response.",
        "properties": {
          "runs": {
            "description": "Newest first, by start time; runs with no manifest sort last, since\nnothing on disk says when they began.",
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array"
          }
        },
        "required": [
          "runs"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "email": "arturh@arturh.de",
      "name": "Artur Hallmann"
    },
    "description": "factorio-bot HTTP API",
    "license": {
      "name": ""
    },
    "title": "factorio-bot",
    "version": "0.2.4-dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/frames": {
      "get": {
        "description": "Derived from a directory listing on every call -- see the module docs for\nwhy this is not computed from a start tick and a stride.",
        "operationId": "list_frames",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FramesManifest"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Returns the manifest of frames captured so far.",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/frames/{client}/{name}": {
      "get": {
        "description": "The route carries no client segment, so `name` is tried against each\ndiscovered `client<N>` directory in turn (lowest client number first),\nanswering the first match. Each directory is bounds-checked with\n[`resolve_script_path`] -- the same traversal guard\n`manage::scripts` uses -- rather than a second, independent\nimplementation of it: this endpoint is unauthenticated, so that guard is\nwhat stands between an HTTP caller and the filesystem.\n\n**Addressed by `(client, name)`, not by `name` alone.** Per-bot cameras mean\n`client1` and `client2` both capture at tick 300 and both files are correct\nand different; they are distinguished by the directory they are in, which\nis exactly what [`FrameEntry::client`] reports. An earlier version of this\nroute searched the client directories in order and returned the first\nmatch, which made the manifest honest and the bytes not: two entries with\nthe same `name` resolved to one image, and the higher-numbered client's\nframe was unreachable. A scrubber showing two bots would have shown the\nsame picture twice and nothing would have said so.\n\nStill no uniquifier: nothing here renames a colliding file so both survive.\nThat would let a genuine same-client double-write live on disk\nindistinguishably from two legitimate frames. The two cases are different —\ntwo clients at one tick is expected and addressable; one client twice at\none tick is a defect and must not be given a home.",
        "operationId": "get_frame",
        "parameters": [
          {
            "description": "the client number, as reported by FrameEntry.client",
            "in": "path",
            "name": "client",
            "required": true,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "a frame's filename, as reported by GET /api/v1/frames",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {}
            },
            "description": "the frame's JPEG bytes"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Serves one frame's JPEG bytes.",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/runs": {
      "get": {
        "operationId": "list_runs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunsResponse"
                }
              }
            },
            "description": "archived runs, newest first"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Lists archived runs.",
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/v1/runs/{id}": {
      "get": {
        "operationId": "get_run",
        "parameters": [
          {
            "description": "the run id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "One run's summary and splits.",
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/v1/runs/{id}/events": {
      "get": {
        "operationId": "get_run_events",
        "parameters": [
          {
            "description": "the run id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Return only events of this `kind`.",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "A run's event log.",
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/v1/runs/{id}/frames": {
      "get": {
        "operationId": "get_run_frames",
        "parameters": [
          {
            "description": "the run id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunFramesResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "A run's frame index.",
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/v1/runs/{id}/frames/{bot}/{name}": {
      "get": {
        "operationId": "get_run_frame",
        "parameters": [
          {
            "description": "the run id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "the bot number, as reported by ArchivedFrame.bot",
            "in": "path",
            "name": "bot",
            "required": true,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "the frame filename, as reported by GET /api/v1/runs/{id}/frames",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {}
            },
            "description": "the frame's JPEG bytes"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "One archived frame's bytes.",
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/v1/runs/{id}/lanes": {
      "get": {
        "operationId": "get_run_lanes",
        "parameters": [
          {
            "description": "the run id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunLanesResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "What each bot did, derived from the run's event log.",
        "tags": [
          "Runs"
        ]
      }
    },
    "/api/v1/runs/{id}/map": {
      "get": {
        "operationId": "get_run_map",
        "parameters": [
          {
            "description": "the run id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunMapResponse"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "A run's entity map: what got built, and whether the game agreed.",
        "tags": [
          "Runs"
        ]
      }
    },
...
  },
...
}